text-ml-model/requirements.txt
annotated-types==0.6.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
dateparser==1.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
idna==3.6
Jinja2==3.1.3
langcodes==3.3.0
MarkupSafe==2.1.5
murmurhash==1.0.10
numpy==1.26.4
packaging==23.2
preshed==3.0.9
pydantic==2.6.2
pydantic_core==2.16.3
python-dateutil==2.8.2
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.3
tqdm==4.66.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
